Cleanup: ensure all coroutines use fnRunning
This greatly simplifies global running flags as they are all moved into
the one fnRunning system. Deadlocks are still possible and in cases
where fnRunning is already started this needs to be communicated to the
function that it should or should not be allowed to wait for a previous
fnRunning to complete.